requirements.txt
altair==5.4.1
altgraph==0.17.4
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
Django==5.1.2
docutils==0.21.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozendict==2.4.6
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
html5lib==1.1
humanfriendly==10.0
idna==3.10
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyboard==0.13.5
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.10
narwhals==1.10.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.0
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
peewee==3.17.7
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
protobuf==5.28.3
pyarrow==17.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMatting==1.1.12
PyMsgBox==1.0.9
pyparsing==3.1.4
pyperclip==1.9.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pystray==0.19.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
reportlab==4.2.5
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
schedule==1.2.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.25.0
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
streamlit==1.39.0
sympy==1.13.3
tenacity==9.0.0
textblob==0.18.0.post0
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2024.9.20
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
ttkthemes==3.2.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
watchdog==5.0.3
weasel==0.4.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
yfinance==0.2.49
